// packages/react-naver-maps/src/types/shape-style.ts
export interface StrokeStyle {
  strokeWeight?: number;
  strokeOpacity?: number;
  strokeColor?: string;
  strokeStyle?: naver.maps.StrokeStyleType;
  strokeLineCap?: naver.maps.StrokeLineCapType;
  strokeLineJoin?: naver.maps.StrokeLineJoinType;
}

export interface FillStyle {
  fillColor?: string;
  fillOpacity?: number;
}